Refactor towards a hexagonal architecture style
Not sure if this is actually better, but the Core Lib is almost free of technical dependencies. Also, it should be easier now to replace the implementation of the calls to the Gitlab API or swap out the persistence.
There are some open questions in my mind:
* should the `FromJSON` instances for the data types live outside the core module as it is a technical detail? This would mean that these instances are orphan instances (how bad would that be?)
* The fine grained `Has*` type classes are maybe redundant now. If so, they should be removed

More discussion should be done in #3