Make the calls to the Gitlab API more resilient
Using `httpJSONEither` avoids throwing an exception when there's an error on JSON conversion. This should mean that the thread for updating statuses doesn't crash when this happens. Instead, the exception will be logged and the other projects are evaluated as normal.
We should kinda expect errors in JSON conversion as the API might return some values that aren't documented. See 6e1ee04