If a pipeline is successful, double check it via the Pipeline endpoint
A pipeline can pass with warnings. This is not reflected in the API when getting the pipelines for a project (the status of the pipeline is `success` in this case). The endpoint for getting a single pipeline contains a block called `detailed_status` which is different according to whether or not there were warnings.

This is a workaround, ideally the request for a projects' pipelines should contain the "correct" status. This might come in a future version of the Gitlab API, see: https://gitlab.com/gitlab-org/gitlab/-/issues/229137