Don't show the runner's name on the UI
It just doesn't help. If you configure a runner's description, it will
show up as the name in the toml config file on the runner. If you try to
configure name *and* description, it won't even start up. The value of
the name field in the toml config will appear as description in the API
response. I haven't found out yet how to change the value of the name
field in the API