Improve error handling for HTTP requests
Previously (due to `httpJSONEither`) all response bodies were parsed as
JSON. This meant that in case of failed requests (read "everything that
doesn't return a 200 OK response"), the overall call failed with a JSON
conversion error. While this is technically correct, it's always hard to
find out what the problem was.
This is now improved. There's an explicit case in `UpdateError` now to
cover that, and `ConversionError` is changed as well. Because `httpLBS`
is now used for doing the actual requests, there's no `JSONException`
anymore. `JSONError` steals the idea of `JSONException` to include
request and reponse in addition to the actual error.
It's a bit more code, but for the consumer it should turn out better
than before