Only query runners with group_type
Otherwise, shared runners will be included as well. This is actually not
that bad (and might be something that should be configurable), but the
tokens of most (all?) users aren't allowed to get the jobs of these
runners which spams the log and doesn't help anyone.
This is rather a workaround than a proper solution, but I haven't found
the time to really dig deeper into that