Show the number of waiting jobs on the runners view
This requires a new concurrent job that regularly fetches all the data
necessary and persists it in an additional IORef.
Since the job API only allows fetching jobs for a given project (and not
for all projects in a group with one query), the new job needs the
project cache as well to run. Therefore, I moved both the projects and
the runners cache into the backbone.
Unfortunately, the jobs representation returned by this endpoint is
different to what is returned in the jobs endpoint for a CI runner, so I
had to add a dedicated data structure for it.
For now, jobs with the status being one of Created, Pending, Preparing,
Scheduled, WaitingForResource is considered waiting.
Since the information is shown on the runners view, the job will only be
activated when the runners view is enabled.

I also went ahead and made the project compile with the --pedantic flag,
that's why there are a whole lot of new pragmas disabling orphan
instance warnings.
The rest of the changes was made to facilitate the reuse of code.