Use a `Reader` in the HTML monad to pass the frontend state around
The reason to make the switch from blaze to lucid in the previous commit
was to make this change easy. By using `Reader`, the arguments don't
have to be threaded through all the functions explicitly, without losing
the flexibility of the monadic HTML DSL. blaze doesn't support that.

Building the HTML representation is still pure, the state that's
required to keep the frontend sane is still completely represented in
the URL, and it's now more clear because of the explicit data type.
It might be necessary to pass the frontend state around even in the
builds view to be able to link back properly, but since there's no link
back at the moment, I didn't care about that in this commit.