Introduce a first version of a UI mode that increases accessibility
For now, it mostly addresses users with color vision deficiencies.
Previously, users have to use the color of the different tiles to
determine the state of a pipeline, which is impossible for users
with color blindness.
The new mode can be accessed via a link on the page or by appending
`?ui=no-color` to the URL. The state of each pipeline will appear as
text below the pipeline name. In addition to that, the background of the
tiles contain a checkmark or a cross to indicate whether a pipeline is
healthy or not.

That's only the first iteration, in the future the positioning of the
icons can be improved, there could be an icon for running pipelines, and
there is a lot of room for refactorings behind the scenes.

This addresses #185