Fetch the statuses for the latest pipelines of all active schedules
When fetching the latest pipeline for each project, scheduled pipelines
are excluded. In addition to that, all active schedules are fetched for
each project, and the latest pipeline for each schedule will be included
in the results.

In the UI, the tiles for each status now include the branch name (for
non-scheduled pipelines) or the schedule description (for scheduled
pipelines). The user can choose whether all, only scheduled, or only
non-scheduled pipelines should be displayed.

Refers to #222